# Use a faster crate for ASCII representation parsing