After:

```
process ASCII 110368129515784116
                        time:   [8.4860 ns 8.5083 ns 8.5317 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
```

Before:

```
process ASCII 110368129515784116
                        time:   [10.923 ns 10.946 ns 10.990 ns]
                        change: [+27.961% +28.231% +28.525%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe
```